refactored SolarPanelDTO and SolarPanelInstallationDTO
