+++ backend/src/main/java/com/SunCycle/SunCycle/dto/SolarPanelRequestDTO.java
@@ -1,8 +1,6 @@
 package com.SunCycle.SunCycle.dto;
 
-import com.SunCycle.SunCycle.model.SolarPanel;
-
-public class SolarPanelDTO {
+public class SolarPanelRequestDTO {
 
     // request fields
     private int modelId;
@@ -11,38 +9,18 @@ public class SolarPanelDTO {
     private String retirementDate;
     private String recyclingMethod;
 
-    // response fields
-    private SolarPanel solarPanel;
-    private String message;
-    private Status status;
 
     // default constructor
-    public SolarPanelDTO() {}
+    public SolarPanelRequestDTO() {}
 
     // constructors for request data
-    public SolarPanelDTO(int modelId, int installationId, String installationDate,
-                         String retirementDate, String recyclingMethod) {
+    public SolarPanelRequestDTO(int modelId, int installationId, String installationDate,
+                                String retirementDate, String recyclingMethod) {
         this.modelId = modelId;
         this.installationId = installationId;
         this.installationDate = installationDate;
         this.retirementDate = retirementDate;
         this.recyclingMethod = recyclingMethod;
-    }
-
-    // constructors for response data
-    public SolarPanelDTO(SolarPanel solarPanel, String message) {
-        this.solarPanel = solarPanel;
-        this.message = message;
-    }
-
-    public SolarPanelDTO(SolarPanel solarPanel, Status status) {
-        this.solarPanel = solarPanel;
-        this.status = status;
-    }
-
-    public SolarPanelDTO(String message, Status status) {
-        this.message = message;
-        this.status = status;
     }
 
     // getters and setters for request fields
@@ -86,29 +64,5 @@ public class SolarPanelDTO {
         this.recyclingMethod = recyclingMethod;
     }
 
-    // getters and setters for response fields
-    public SolarPanel getSolarPanel() {
-        return solarPanel;
-    }
-
-    public void setSolarPanel(SolarPanel solarPanel) {
-        this.solarPanel = solarPanel;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public Status getStatus() {
-        return status;
-    }
-
-    public void setStatus(Status status) {
-        this.status = status;
-    }
 
 }
