+++ backend/src/main/java/com/SunCycle/SunCycle/service/SolarPanelService.java
@@ -1,6 +1,7 @@
 package com.SunCycle.SunCycle.service;
 
-import com.SunCycle.SunCycle.dto.SolarPanelDTO;
+import com.SunCycle.SunCycle.dto.SolarPanelRequestDTO;
+import com.SunCycle.SunCycle.dto.SolarPanelResponseDTO;
 import com.SunCycle.SunCycle.dto.Status;
 import com.SunCycle.SunCycle.model.*;
 import com.SunCycle.SunCycle.repository.SolarPanelInstallationRepository;
@@ -23,7 +24,7 @@ public class SolarPanelService {
     @Autowired
     private SolarPanelInstallationRepository solarPanelInstallationRepository;
 
-    public SolarPanelDTO createPanel(SolarPanelDTO dto) {
+    public SolarPanelResponseDTO createPanel(SolarPanelRequestDTO dto) {
         // convert dto to panel instance
         SolarPanel panel = new SolarPanel();
 
@@ -32,13 +33,13 @@ public class SolarPanelService {
         if (model.isPresent())
             panel.setModel(model.get());
         else
-            return new SolarPanelDTO("Panel model not found", Status.NOT_FOUND);
+            return new SolarPanelResponseDTO("Panel model not found", Status.NOT_FOUND);
 
         Optional<SolarPanelInstallation> installation = solarPanelInstallationRepository.findById(dto.getInstallationId());
         if (installation.isPresent())
             panel.setInstallation(installation.get());
         else
-            return new SolarPanelDTO("Installation not found", Status.NOT_FOUND);
+            return new SolarPanelResponseDTO("Installation not found", Status.NOT_FOUND);
 
         // set other fields
         panel.setInstallationDate(dto.getInstallationDate());
@@ -46,15 +47,15 @@ public class SolarPanelService {
         panel.setRecyclingMethod(dto.getRecyclingMethod());
 
         // return response DTO
-        return new SolarPanelDTO(solarPanelRepository.save(panel), Status.SUCCESS);
+        return new SolarPanelResponseDTO(solarPanelRepository.save(panel), Status.SUCCESS);
     }
 
-    public SolarPanelDTO updatePanel(int panelId, SolarPanelDTO dto) {
+    public SolarPanelResponseDTO updatePanel(int panelId, SolarPanelRequestDTO dto) {
         // try to fetch panel by given id
         Optional<SolarPanel> panelOpt = solarPanelRepository.findById(panelId);
 
         if (panelOpt.isEmpty())
-            return new SolarPanelDTO("Panel not found", Status.NOT_FOUND);
+            return new SolarPanelResponseDTO("Panel not found", Status.NOT_FOUND);
 
         SolarPanel panel = panelOpt.get();
 
@@ -62,13 +63,13 @@ public class SolarPanelService {
         // check if new model exists
         Optional<SolarPanelModel> model = solarPanelModelRepository.findById(dto.getModelId());
         if (model.isEmpty())
-            return new SolarPanelDTO("Model not found", Status.NOT_FOUND);
+            return new SolarPanelResponseDTO("Model not found", Status.NOT_FOUND);
         panel.setModel(model.get());
 
         // check if new installation exists
         Optional<SolarPanelInstallation> installation = solarPanelInstallationRepository.findById(dto.getInstallationId());
         if (installation.isEmpty())
-            return new SolarPanelDTO("Installation not found", Status.NOT_FOUND);
+            return new SolarPanelResponseDTO("Installation not found", Status.NOT_FOUND);
         panel.setInstallation(installation.get());
 
         // update the rest fields
@@ -76,19 +77,19 @@ public class SolarPanelService {
         panel.setRetirementDate(dto.getRetirementDate());
         panel.setRecyclingMethod(dto.getRecyclingMethod());
 
-        return new SolarPanelDTO(solarPanelRepository.save(panel), Status.SUCCESS);
+        return new SolarPanelResponseDTO(solarPanelRepository.save(panel), Status.SUCCESS);
     }
 
-    public SolarPanelDTO deletePanel(int panelId) {
+    public SolarPanelResponseDTO deletePanel(int panelId) {
         // try to fetch panel by given id
         Optional<SolarPanel> panelOpt = solarPanelRepository.findById(panelId);
 
         if (panelOpt.isEmpty())
-            return new SolarPanelDTO("Panel not found", Status.NOT_FOUND);
+            return new SolarPanelResponseDTO("Panel not found", Status.NOT_FOUND);
 
         solarPanelRepository.delete(panelOpt.get());
 
-        return new SolarPanelDTO("Panel deleted successfully", Status.SUCCESS);
+        return new SolarPanelResponseDTO("Panel deleted successfully", Status.SUCCESS);
     }
 
 }
